refactor(cli): split CLITuiMixin into topical siblings
Keep cli_tui_mixin.py as the public facade and move its 104 methods into
panel, chrome, handler, submission, and construction siblings. Method bodies
are mechanically preserved and the facade MRO keeps every method reachable.

Replace the two source-reading tests that blocked this cleanup with focused
behavioral tests for inline repaint and type-ahead routing. Update the CLI
extension guide to name the construction sibling.

Evidence: 108/108 method source segments byte-identical (including helper
methods), py_compile/import/MRO checks green, focused CLI tests 31/31 green,
and full tests/cli/ 1424 passed / 2 failed / 8 skipped; the two failures are
identical on untouched main (pre-existing focus-view failures).
