+++ website/docs/developer-guide/extending-the-cli.md
@@ -6,7 +6,7 @@ description: "Build wrapper CLIs that extend the Hermes TUI with custom widgets,
 
 # Extending the CLI
 
-Hermes exposes protected extension hooks on `HermesCLI` so wrapper CLIs can add widgets, keybindings, and layout customizations without overriding the `run()` method or the TUI construction in `hermes_cli/cli_tui_mixin.py` (where these hooks are defined; `HermesCLI` in `cli.py` mixes it in). This keeps your extension decoupled from internal changes.
+Hermes exposes protected extension hooks on `HermesCLI` so wrapper CLIs can add widgets, keybindings, and layout customizations without overriding the `run()` method or the TUI construction in `hermes_cli/cli_tui_mixin.py` (facade; hooks remain there, while implementation topics live in `hermes_cli/cli_tui_construction.py` and related siblings). This keeps your extension decoupled from internal changes.
 
 ## Extension points
